Feat: atualização da sidebar
