+++ supermercadosq/src/components/PasswordSecurity/PasswordSecurity.jsx
@@ -1,13 +1,9 @@
 import React from 'react';
-import { SideLink } from './styles'
 
 
 export default function PasswordSecurity() {
   return (
      <>
-      <SideLink to='/login/profile/teste3'>
-        Senhas e segurança
-      </SideLink>
      </>
   );
 }
