+++ supermercadosq/src/components/PostComments/PostComments.jsx
@@ -4,9 +4,7 @@ import { SideLink } from './styles'
 export default function PostComments() {
   return (
      <>
-      <SideLink to='/login/profile/teste2'>
-        Postagens e comentários
-      </SideLink>
+      
      </>
   );
 }
