--- supermercadosq/src/components/ProfileInformation/styles.jsx
@@ -1,14 +1,0 @@
-import styled from 'styled-components';
-import { Link } from 'react-router-dom';
-
-export const SideLink = styled(Link)`
-  text-decoration: none;
-  color: rgba(144, 144, 144, 1);
-  font-weight: bold;
-
-  .icone {
-    font-size: 2rem;
-
-  }
-  
-`;
