+++ supermercadosq/src/components/SideBar/SideBar.jsx
@@ -1,15 +1,30 @@
 import React from 'react';
 import { SideBarContainer } from './styles'
-import InformacoesPerfil from '../../components/ProfileInformation/ProfileInformation';
-import SenhasSegurança from '../../components/PasswordSecurity/PasswordSecurity';
-import PostagensComentarios from '../../components/PostComments/PostComments';
+import { SideLink } from './styles';
+import { BiChevronRight } from 'react-icons/bi';
+
 
 export default function SideBar() {
   return (
     <SideBarContainer>
-      <InformacoesPerfil />
-      <SenhasSegurança />
-      <PostagensComentarios />
+       <SideLink to='/login/profile'>
+        <div style={{ display: 'flex', alignItems: 'center'}} >
+          Informações do perfil           
+          <BiChevronRight size='1.5rem' />
+        </div>  
+      </SideLink>
+
+      <SideLink to='/login/profile/teste2' >
+        <div style={{ display: 'flex', alignItems: 'center', margin: '1rem 0'}}>
+          Senhas e segurança 
+        </div>  
+      </SideLink>
+
+      <SideLink to='/login/profile/teste3' >
+        <div style={{ display: 'flex', alignItems: 'center'}}>
+          Postagens e comentários
+        </div>  
+      </SideLink>
     </SideBarContainer>
   );
 }
